parser-json-shchk: separate module for ShellCheckTreeDecoder
